release 0.0.65 source code for go

+++ services/bcs/v2/model/model_basic_info.go
@@ -19,21 +19,21 @@ type BasicInfo struct {
 	// BCS服务的类型，分为：联盟链（union），私有链（private）
 
 	ServiceType *string `json:"service_type,omitempty"`
-	// BCS服务部署类型，分为一键购买（onestep），普通购买（normal）
+	// BCS服务部署类型，一键部署（onestep），普通部署（normal）
 
 	PurchaseType *string `json:"purchase_type,omitempty"`
 	// BCS服务安全机制，分为ECDSA（ECDSA），国密算法（sm2）
 
 	SignAlgorithm *string `json:"sign_algorithm,omitempty"`
-	// BCS服务的共识策略，分为测试策略（solo），快速拜占庭容错算法（sflic）,Kafka(kafka)
+	// BCS服务的共识策略，分为测试策略（solo），快速拜占庭容错算法（sflic）,Kafka(kafka)，raft共识算法（etcdraft）
 
 	Consensus *string `json:"consensus,omitempty"`
-	// BCS服务付费模式，分为按需（1），包周期（0）
+	// BCS服务付费模式，分为按需（1）[包周期（0）](tag:onorder)
 
-	ChargingMode *int32 `json:"charging_mode,omitempty"`
-	// BCS服务版本类型，分为基础版（1），专业版（2），铂金版（3）
+	ChargingMode *int64 `json:"charging_mode,omitempty"`
+	// BCS服务版本类型
 
-	VersionType *int32 `json:"version_type,omitempty"`
+	VersionType *int64 `json:"version_type,omitempty"`
 	// BCS服务数据库类型，包括文件数据库（goleveldb），NoSQL（couchdb）
 
 	DatabaseType *string `json:"database_type,omitempty"`
@@ -43,7 +43,7 @@ type BasicInfo struct {
 	// BCS服务所在集群名称
 
 	ClusterName *string `json:"cluster_name,omitempty"`
-	// BCS服务的集群类型，分为CCE集群（空），IEF集群（ief）
+	// BCS服务的集群类型，分为CCE集群（CCE），IEF集群（ief）
 
 	ClusterType *string `json:"cluster_type,omitempty"`
 	// BCS多可用区标示，分为：多可用区（yes），非多可用区（no）
@@ -81,7 +81,7 @@ type BasicInfo struct {
 	ProcessStatus *string `json:"process_status,omitempty"`
 	// BCS服务为包周期模式时，返回值为0（订单未成功）,1（订单异常）,2（订单正常）
 
-	OrderStatus *int32 `json:"order_status,omitempty"`
+	OrderStatus *int64 `json:"order_status,omitempty"`
 }
 
 func (o BasicInfo) String() string {
