+++ services/bcs/v2/model/model_create_request_body_block_info.go
@@ -10,13 +10,13 @@ import (
 type CreateRequestBodyBlockInfo struct {
 	// 区块产生时间（单位：秒），默认2秒
 
-	BatchTimeout *int32 `json:"batch_timeout,omitempty"`
+	BatchTimeout *int64 `json:"batch_timeout,omitempty"`
 	// 区块包含交易数量，默认500
 
-	MaxMessageCount *int32 `json:"max_message_count,omitempty"`
+	MaxMessageCount *int64 `json:"max_message_count,omitempty"`
 	// 区块容量（单位：MB），默认2MB
 
-	PreferredMaxbytes *int32 `json:"preferred_maxbytes,omitempty"`
+	PreferredMaxbytes *int64 `json:"preferred_maxbytes,omitempty"`
 }
 
 func (o CreateRequestBodyBlockInfo) String() string {
