+++ services/bcs/v2/model/model_resource.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Resource struct {
-	// 资源类型，包含instance（服务实例数）和peers（总peer数）两种
+	// 资源类型，包含instance（服务实例数）
 
 	Type *string `json:"type,omitempty"`
 	// 单位
@@ -15,19 +15,19 @@ type Resource struct {
 	Unit *string `json:"unit,omitempty"`
 	// 最小值
 
-	Min *int32 `json:"min,omitempty"`
+	Min *int64 `json:"min,omitempty"`
 	// 最大值
 
-	Max *int32 `json:"max,omitempty"`
+	Max *int64 `json:"max,omitempty"`
 	// 配额
 
-	Quota *int32 `json:"quota,omitempty"`
+	Quota *int64 `json:"quota,omitempty"`
 	// 已使用配额
 
-	Used *int32 `json:"used,omitempty"`
+	Used *int64 `json:"used,omitempty"`
 	// 剩余配额
 
-	Free *int32 `json:"free,omitempty"`
+	Free *int64 `json:"free,omitempty"`
 }
 
 func (o Resource) String() string {
